pyproject.toml
[build-system]
requires = [
    "setuptools==70.1.1",
    "wheel==0.43.0"
]
build-backend = "setuptools.build_meta"

[project]
name = "sso_manager"
version = "0.1.0"
description = "A brief description of your project"
authors = [
    { name = "Islam Elkadi", email = "ielkadi@icloud.com" }
]
dependencies = [
    "jsonschema==4.22.0",
    "boto3==1.34.136",
    "PyYAML==6.0.2"
]

[project.optional-dependencies]
dev = [
    "pre-commit==3.6.0",
    "pyclean==3.0.0",
    "moto==5.0.11",
    "pytest==8.1.1",
    "pytest-xdist==3.6.1",
    "pytest-cov==5.0.0",
    "pylint==3.0.3",
    "black==23.12.0",
]  # Add any development dependencies here

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v -ra -q -p no:warnings"
log_cli = true
log_cli_level = "INFO"
log_format = "%(asctime)s %(levelname)s %(message)s"
log_date_format = "%Y-%m-%d %H:%M:%S"
minversion = "6.0"
filterwarnings = "ignore"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
